PM-42388: chore: Separate Cipher flows from VaultRepository
