+++ app/src/main/kotlin/com/x8bit/bitwarden/data/vault/repository/VaultRepositoryImpl.kt
@@ -1,24 +1,16 @@
 package com.x8bit.bitwarden.data.vault.repository
 
 import com.bitwarden.core.InitUserCryptoMethod
-import com.bitwarden.core.data.manager.dispatcher.DispatcherManager
 import com.bitwarden.core.data.manager.model.FlagKey
 import com.bitwarden.core.data.repository.error.MissingPropertyException
-import com.bitwarden.core.data.repository.model.DataState
 import com.bitwarden.core.data.repository.util.bufferedMutableSharedFlow
-import com.bitwarden.core.data.repository.util.combineDataStates
-import com.bitwarden.core.data.repository.util.map
-import com.bitwarden.core.data.repository.util.mapNullable
 import com.bitwarden.core.data.util.asFailure
 import com.bitwarden.exporters.ExportFormat
 import com.bitwarden.fido.Fido2CredentialAutofillView
 import com.bitwarden.sdk.Fido2CredentialStore
-import com.bitwarden.send.SendView
 import com.bitwarden.vault.CipherListView
 import com.bitwarden.vault.CipherListViewType
 import com.bitwarden.vault.CipherType
-import com.bitwarden.vault.CipherView
-import com.bitwarden.vault.FolderView
 import com.x8bit.bitwarden.data.auth.datasource.disk.AuthDiskSource
 import com.x8bit.bitwarden.data.auth.repository.util.toSdkParams
 import com.x8bit.bitwarden.data.autofill.util.login
@@ -32,12 +24,10 @@ import com.x8bit.bitwarden.data.vault.manager.CredentialExchangeImportManager
 import com.x8bit.bitwarden.data.vault.manager.FolderManager
 import com.x8bit.bitwarden.data.vault.manager.PinProtectedUserKeyManager
 import com.x8bit.bitwarden.data.vault.manager.SendManager
-import com.x8bit.bitwarden.data.vault.manager.TotpCodeManager
+import com.x8bit.bitwarden.data.vault.manager.VaultDataManager
 import com.x8bit.bitwarden.data.vault.manager.VaultLockManager
 import com.x8bit.bitwarden.data.vault.manager.VaultSyncManager
-import com.x8bit.bitwarden.data.vault.manager.model.GetCipherResult
 import com.x8bit.bitwarden.data.vault.manager.model.ImportCxfPayloadResult
-import com.x8bit.bitwarden.data.vault.manager.model.VerificationCodeItem
 import com.x8bit.bitwarden.data.vault.repository.model.ExportVaultDataResult
 import com.x8bit.bitwarden.data.vault.repository.model.GenerateTotpResult
 import com.x8bit.bitwarden.data.vault.repository.model.ImportCredentialsResult
@@ -48,21 +38,9 @@ import com.x8bit.bitwarden.data.vault.repository.util.toEncryptedSdkCipher
 import com.x8bit.bitwarden.data.vault.repository.util.toEncryptedSdkFolder
 import com.x8bit.bitwarden.data.vault.repository.util.toSdkAccount
 import com.x8bit.bitwarden.data.vault.repository.util.toSdkMasterPasswordUnlock
-import com.x8bit.bitwarden.ui.vault.feature.vault.model.VaultFilterType
-import com.x8bit.bitwarden.ui.vault.feature.vault.util.toFilteredList
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.firstOrNull
-import kotlinx.coroutines.flow.flatMapLatest
-import kotlinx.coroutines.flow.flowOf
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.launch
 import timber.log.Timber
 import java.security.GeneralSecurityException
 import java.time.Instant
@@ -80,179 +58,24 @@ internal class VaultRepositoryImpl(
     private val folderManager: FolderManager,
     private val sendManager: SendManager,
     private val vaultLockManager: VaultLockManager,
-    private val totpCodeManager: TotpCodeManager,
     private val vaultSyncManager: VaultSyncManager,
     private val credentialExchangeImportManager: CredentialExchangeImportManager,
     private val pinProtectedUserKeyManager: PinProtectedUserKeyManager,
     private val featureFlagManager: FeatureFlagManager,
-    dispatcherManager: DispatcherManager,
+    vaultDataManager: VaultDataManager,
 ) : VaultRepository,
     CipherManager by cipherManager,
     FolderManager by folderManager,
     SendManager by sendManager,
+    VaultDataManager by vaultDataManager,
     VaultLockManager by vaultLockManager,
     VaultSyncManager by vaultSyncManager {
-
-    private val unconfinedScope = CoroutineScope(dispatcherManager.unconfined)
-    private val ioScope = CoroutineScope(dispatcherManager.io)
-
     private val activeUserId: String? get() = authDiskSource.userState?.activeUserId
 
     private val mutableTotpCodeResultFlow = bufferedMutableSharedFlow<TotpCodeResult>()
 
-    override var vaultFilterType: VaultFilterType = VaultFilterType.AllVaults
-
     override val totpCodeFlow: Flow<TotpCodeResult>
         get() = mutableTotpCodeResultFlow.asSharedFlow()
-
-    override fun deleteVaultData(userId: String) {
-        ioScope.launch {
-            vaultDiskSource.deleteVaultData(userId)
-        }
-    }
-
-    override fun getVaultItemStateFlow(itemId: String): StateFlow<DataState<CipherView?>> =
-        vaultSyncManager
-            .vaultDataStateFlow
-            .map { dataState ->
-                dataState.map { vaultData ->
-                    val getCipherResult = vaultData
-                        .decryptCipherListResult
-                        .successes
-                        .find { it.id == itemId }
-                        .let { getCipher(itemId) }
-                    when (getCipherResult) {
-                        is GetCipherResult.Success -> getCipherResult.cipherView
-                        else -> null
-                    }
-                }
-            }
-            .stateIn(
-                scope = unconfinedScope,
-                started = SharingStarted.Lazily,
-                initialValue = DataState.Loading,
-            )
-
-    override fun getVaultListItemStateFlow(itemId: String): StateFlow<DataState<CipherListView?>> =
-        vaultSyncManager
-            .vaultDataStateFlow
-            .map { dataState ->
-                dataState.map { vaultData ->
-                    vaultData
-                        .decryptCipherListResult
-                        .successes
-                        .find { it.id == itemId }
-                }
-            }
-            .stateIn(
-                scope = unconfinedScope,
-                started = SharingStarted.Lazily,
-                initialValue = DataState.Loading,
-            )
-
-    override fun getVaultFolderStateFlow(folderId: String): StateFlow<DataState<FolderView?>> =
-        vaultSyncManager
-            .vaultDataStateFlow
-            .map { dataState ->
-                dataState.map { vaultData ->
-                    vaultData
-                        .folderViewList
-                        .find { it.id == folderId }
-                }
-            }
-            .stateIn(
-                scope = unconfinedScope,
-                started = SharingStarted.Lazily,
-                initialValue = DataState.Loading,
-            )
-
-    override fun getSendStateFlow(sendId: String): StateFlow<DataState<SendView?>> =
-        vaultSyncManager
-            .sendDataStateFlow
-            .map { dataState ->
-                dataState.map { sendData ->
-                    sendData.sendViewList.find { it.id == sendId }
-                }
-            }
-            .stateIn(
-                scope = unconfinedScope,
-                started = SharingStarted.Lazily,
-                initialValue = DataState.Loading,
-            )
-
-    @OptIn(ExperimentalCoroutinesApi::class)
-    override fun getAuthCodesFlow(): StateFlow<DataState<List<VerificationCodeItem>>> {
-        val userId = activeUserId ?: return MutableStateFlow(
-            DataState.Error(IllegalStateException("No active user"), null),
-        )
-        return vaultSyncManager
-            .vaultDataStateFlow
-            .map { dataState ->
-                dataState.map { vaultData ->
-                    vaultData
-                        .decryptCipherListResult
-                        .successes
-                        .filter {
-                            it.type is CipherListViewType.Login &&
-                                !it.login?.totp.isNullOrBlank() &&
-                                it.isActive
-                        }
-                        .toFilteredList(vaultFilterType)
-                }
-            }
-            .flatMapLatest { cipherDataState ->
-                val cipherList = cipherDataState.data ?: emptyList()
-                totpCodeManager
-                    .getTotpCodesForCipherListViewsStateFlow(
-                        userId = userId,
-                        cipherListViews = cipherList,
-                    )
-                    .map { verificationCodeDataStates ->
-                        combineDataStates(
-                            verificationCodeDataStates,
-                            cipherDataState,
-                        ) { verificationCodeItems, _ ->
-                            // Just return the verification items; we are only combining the
-                            // DataStates to know the overall state.
-                            verificationCodeItems
-                        }
-                    }
-            }
-            .stateIn(
-                scope = unconfinedScope,
-                started = SharingStarted.WhileSubscribed(),
-                initialValue = DataState.Loading,
-            )
-    }
-
-    @OptIn(ExperimentalCoroutinesApi::class)
-    override fun getAuthCodeFlow(cipherId: String): StateFlow<DataState<VerificationCodeItem?>> {
-        val userId = activeUserId ?: return MutableStateFlow(
-            DataState.Error(IllegalStateException("No active user"), null),
-        )
-        return this.getVaultListItemStateFlow(cipherId)
-            .flatMapLatest { cipherDataState ->
-                cipherDataState
-                    .data
-                    ?.let {
-                        totpCodeManager
-                            .getTotpCodeStateFlow(userId = userId, cipherListView = it)
-                            .map { totpCodeDataState ->
-                                combineDataStates(totpCodeDataState, cipherDataState) { _, _ ->
-                                    // We are only combining the DataStates to know the overall
-                                    // state, we map it to the appropriate value below.
-                                }
-                                    .mapNullable { totpCodeDataState.data }
-                            }
-                    }
-                    ?: flowOf(DataState.Loaded(null))
-            }
-            .stateIn(
-                scope = unconfinedScope,
-                started = SharingStarted.WhileSubscribed(),
-                initialValue = DataState.Loading,
-            )
-    }
 
     override suspend fun silentlyDiscoverCredentials(
         userId: String,
