+++ app/src/main/kotlin/com/x8bit/bitwarden/data/vault/repository/di/VaultRepositoryModule.kt
@@ -1,6 +1,5 @@
 package com.x8bit.bitwarden.data.vault.repository.di
 
-import com.bitwarden.core.data.manager.dispatcher.DispatcherManager
 import com.x8bit.bitwarden.data.auth.datasource.disk.AuthDiskSource
 import com.x8bit.bitwarden.data.platform.manager.FeatureFlagManager
 import com.x8bit.bitwarden.data.vault.datasource.disk.VaultDiskSource
@@ -10,7 +9,7 @@ import com.x8bit.bitwarden.data.vault.manager.CredentialExchangeImportManager
 import com.x8bit.bitwarden.data.vault.manager.FolderManager
 import com.x8bit.bitwarden.data.vault.manager.PinProtectedUserKeyManager
 import com.x8bit.bitwarden.data.vault.manager.SendManager
-import com.x8bit.bitwarden.data.vault.manager.TotpCodeManager
+import com.x8bit.bitwarden.data.vault.manager.VaultDataManager
 import com.x8bit.bitwarden.data.vault.manager.VaultLockManager
 import com.x8bit.bitwarden.data.vault.manager.VaultSyncManager
 import com.x8bit.bitwarden.data.vault.repository.VaultRepository
@@ -38,11 +37,10 @@ object VaultRepositoryModule {
         folderManager: FolderManager,
         sendManager: SendManager,
         vaultLockManager: VaultLockManager,
-        dispatcherManager: DispatcherManager,
-        totpCodeManager: TotpCodeManager,
         vaultSyncManager: VaultSyncManager,
         credentialExchangeImportManager: CredentialExchangeImportManager,
         pinProtectedUserKeyManager: PinProtectedUserKeyManager,
+        vaultDataManager: VaultDataManager,
         featureFlagManager: FeatureFlagManager,
     ): VaultRepository = VaultRepositoryImpl(
         vaultDiskSource = vaultDiskSource,
@@ -52,11 +50,10 @@ object VaultRepositoryModule {
         folderManager = folderManager,
         sendManager = sendManager,
         vaultLockManager = vaultLockManager,
-        dispatcherManager = dispatcherManager,
-        totpCodeManager = totpCodeManager,
         vaultSyncManager = vaultSyncManager,
         credentialExchangeImportManager = credentialExchangeImportManager,
         pinProtectedUserKeyManager = pinProtectedUserKeyManager,
+        vaultDataManager = vaultDataManager,
         featureFlagManager = featureFlagManager,
     )
 }
